refactor(test): split aggregate test stages into separate files
- Move GivenStage, WhenStage, ExpectStage, and VerifiedStage interfaces to their own files
- Extract EventIterator and ExpectedResult classes to their own files
- Refactor DefaultGivenStage, DefaultWhenStage, DefaultExpectStage, and DefaultVerifiedStage implementations
- Remove Stages.kt file as its contents have been split into separate files
