+++ test/wow-test/src/main/kotlin/me/ahoo/wow/test/aggregate/ExpectedResult.kt
@@ -1,0 +1,27 @@
+/*
+ * Copyright [2021-present] [ahoo wang <ahoowang@qq.com> (https://github.com/Ahoo-Wang)].
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package me.ahoo.wow.test.aggregate
+
+import me.ahoo.wow.command.ServerCommandExchange
+import me.ahoo.wow.event.DomainEventStream
+import me.ahoo.wow.modeling.state.StateAggregate
+
+data class ExpectedResult<S : Any>(
+    val exchange: ServerCommandExchange<*>,
+    val stateAggregate: StateAggregate<S>,
+    val domainEventStream: DomainEventStream? = null,
+    val error: Throwable? = null
+) {
+    val hasError = error != null
+}
